python-backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
pymongo==4.6.0
motor==3.3.2

# AWS Dependencies
boto3==1.34.0
botocore==1.34.0

# Authentication - Updated for compatibility
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# HTTP client and utilities
httpx==0.25.2
requests==2.31.0

# Date and time utilities
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For PostgreSQL support
# psycopg2-binary==2.9.9

# Optional: For Redis caching
# redis==5.0.1
